Major changes to search algorithm.
The beam partial derivation weights are now calculated according to the
generalized prefix conditional probability of the derivation. This means
that the scores of the partial derivations on the beam can change as the
search progresses, so it no longer makes sense to store the beam as a
priority queue, though I haven't made that change yet.