Bugfix: bug in add_to_pqueue fixed.
-  The bug was
that if an a really bad derivation was inserted into a saturated
pqueue, then the infinum was not being updated when a better
derivation was found. This meant that insertion of derivations was not
being pruned in some cases. This bug should have only affected the
runtime of the code, not the accuracy, but since the code is
time-bound, it's hard to tell.

- Fixed by always setting the infinum to the last element of the
  pqueue (could be expensive for large pqueue).